(REF): Stdlib injection refactoring
